[#76500190] Added endpoint /users/1.0/get (code remarks)
